AYS-523 | `common.util.exception` Package Has Been Moved to `common` Package
