+++ src/main/java/org/ays/auth/util/exception/AysAdminRegistrationApplicationAlreadyApprovedException.java
@@ -1,7 +1,7 @@
 package org.ays.auth.util.exception;
 
 import org.ays.auth.model.enums.AdminRegistrationApplicationStatus;
-import org.ays.common.util.exception.AysAlreadyException;
+import org.ays.common.exception.AysAlreadyException;
 
 import java.io.Serial;
