+++ src/main/java/org/ays/auth/util/exception/AysInvalidRoleStatusException.java
@@ -1,7 +1,7 @@
 package org.ays.auth.util.exception;
 
 import org.ays.auth.model.enums.AysRoleStatus;
-import org.ays.common.util.exception.AysNotExistException;
+import org.ays.common.exception.AysNotExistException;
 
 import java.io.Serial;
