+++ src/main/java/org/ays/auth/util/exception/AysUserAlreadyExistsByPhoneNumberException.java
@@ -1,7 +1,7 @@
 package org.ays.auth.util.exception;
 
+import org.ays.common.exception.AysAlreadyException;
 import org.ays.common.model.AysPhoneNumber;
-import org.ays.common.util.exception.AysAlreadyException;
 
 import java.io.Serial;
