+++ src/main/java/org/ays/auth/util/exception/AysUserDoesNotAccessPageException.java
@@ -1,7 +1,7 @@
 package org.ays.auth.util.exception;
 
 import org.ays.auth.model.enums.AysSourcePage;
-import org.ays.common.util.exception.AysAuthException;
+import org.ays.common.exception.AysAuthException;
 
 import java.io.Serial;
