+++ src/main/java/org/ays/emergency_application/exception/EmergencyEvacuationApplicationNotExistException.java
@@ -1,6 +1,6 @@
 package org.ays.emergency_application.exception;
 
-import org.ays.common.util.exception.AysNotExistException;
+import org.ays.common.exception.AysNotExistException;
 
 import java.io.Serial;
