+++ src/main/java/org/ays/emergency_application/util/annotation/CheckEmergencyEvacuationApplicationActivityAspect.java
@@ -2,7 +2,7 @@ package org.ays.emergency_application.util.annotation;
 
 import org.aspectj.lang.annotation.Aspect;
 import org.aspectj.lang.annotation.Before;
-import org.ays.common.util.exception.EmergencyEvacuationApplicationEndpointNotActiveException;
+import org.ays.common.exception.EmergencyEvacuationApplicationEndpointNotActiveException;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
