+++ src/main/java/org/ays/institution/exception/AysInstitutionNotExistException.java
@@ -1,6 +1,6 @@
 package org.ays.institution.exception;
 
-import org.ays.common.util.exception.AysNotExistException;
+import org.ays.common.exception.AysNotExistException;
 
 import java.io.Serial;
