+++ src/test/java/org/ays/common/util/exception/handler/GlobalExceptionHandlerTest.java
@@ -5,12 +5,13 @@ import com.fasterxml.jackson.databind.exc.InvalidFormatException;
 import jakarta.validation.ConstraintViolationException;
 import jakarta.validation.constraints.NotNull;
 import org.ays.AysRestControllerTest;
+import org.ays.common.exception.AysAuthException;
+import org.ays.common.exception.AysBadRequestException;
+import org.ays.common.exception.AysForbiddenException;
+import org.ays.common.exception.AysNotExistException;
+import org.ays.common.exception.AysProcessException;
+import org.ays.common.exception.handler.GlobalExceptionHandler;
 import org.ays.common.model.response.AysErrorResponse;
-import org.ays.common.util.exception.AysAuthException;
-import org.ays.common.util.exception.AysBadRequestException;
-import org.ays.common.util.exception.AysForbiddenException;
-import org.ays.common.util.exception.AysNotExistException;
-import org.ays.common.util.exception.AysProcessException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
